Big-Voice-Platform/pyproject.toml
[project]
name = "big-voice-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "chromadb>=1.0.16",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.29.0",
    "gtts>=2.5.4",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.4",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langmem>=0.0.29",
    "langmem-adapter>=0.0.2",
    "openai-agents[litellm]>=0.2.4",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "pypdf>=5.9.0",
    "sqlalchemy>=2.0.42",
    "stripe>=12.4.0",
]
